Fixed Reproduction Bugs
There was a problem with newly born Specimen being killed off immediately because they didn't eat in the generationalChange() function, so I had to change the approach a bit. Instead of adding them immediately, I create a matrix of outcomes for existing Specimen and then apply them only to existing ones.